Test (succesful)

+++ Core/Inc/PRBS.h
@@ -7,6 +7,6 @@
 
 // Function prototypes
 
-void detectRotations(uint16_t, uint16_t, int16_t*);
+void detectRotations(uint32_t, uint32_t, int32_t*);
 
 #endif /* INC_PRBS_H_ */
